+++ .storybook/manager.ts
@@ -3,7 +3,4 @@ import {themes} from './theme';
 
 addons.setConfig({
     theme: themes.light,
-    sidebar: {
-        collapsedRoots: ['docs'],
-    },
 });