+++ demo/components/Playground.tsx
@@ -1,5 +1,6 @@
 import {type CSSProperties, memo, useCallback, useEffect, useMemo, useState} from 'react';
 
+import type {EmbeddingMode} from '@diplodoc/html-extension';
 import {defaultOptions} from '@diplodoc/transform/lib/sanitize';
 import {Button, DropdownMenu} from '@gravity-ui/uikit';
 
@@ -77,6 +78,8 @@ export type PlaygroundProps = {
     onChangeEditorType?: (mode: MarkdownEditorMode) => void;
     onChangeSplitModeEnabled?: (splitModeEnabled: boolean) => void;
     directiveSyntax?: DirectiveSyntaxValue;
+    disabledHTMLBlockModes?: EmbeddingMode[];
+    disableMarkdownItAttrs?: boolean;
 } & Pick<UseMarkdownEditorProps, 'experimental' | 'wysiwygConfig'> &
     Pick<
         MarkdownEditorViewProps,
@@ -125,6 +128,8 @@ export const Playground = memo<PlaygroundProps>((props) => {
         hidePreviewAfterSubmit,
         experimental,
         directiveSyntax,
+        disabledHTMLBlockModes,
+        disableMarkdownItAttrs,
     } = props;
     const [editorMode, setEditorMode] = useState<MarkdownEditorMode>(initialEditor ?? 'wysiwyg');
     const [mdRaw, setMdRaw] = useState<MarkupString>(initial || '');
@@ -143,9 +148,11 @@ export const Playground = memo<PlaygroundProps>((props) => {
                 breaks={md.breaks}
                 needToSanitizeHtml={sanitizeHtml}
                 plugins={getPlugins({directiveSyntax})}
+                disableMarkdownItAttrs={disableMarkdownItAttrs}
+                htmlRuntimeConfig={{disabledModes: disabledHTMLBlockModes}}
             />
         ),
-        [sanitizeHtml],
+        [sanitizeHtml, disabledHTMLBlockModes, disableMarkdownItAttrs],
     );
 
     const logger = useMemo(() => new Logger2().nested({env: 'playground'}), []);
@@ -157,6 +164,7 @@ export const Playground = memo<PlaygroundProps>((props) => {
             preset: 'full',
             wysiwygConfig: {
                 placeholderOptions: placeholderOptions,
+                disableMarkdownAttrs: disableMarkdownItAttrs,
                 extensions: (builder) => {
                     builder
                         .use(Math, {
@@ -246,6 +254,7 @@ export const Playground = memo<PlaygroundProps>((props) => {
             experimental?.beforeEditorModeChange,
             experimental?.prepareRawMarkup,
             directiveSyntax,
+            disableMarkdownItAttrs,
         ],
     );
 
@@ -303,6 +312,7 @@ export const Playground = memo<PlaygroundProps>((props) => {
                 <MarkdownEditorView
                     autofocus
                     className={className}
+                    qa="demo-md-editor"
                     stickyToolbar={Boolean(stickyToolbar)}
                     toolbarsPreset={toolbarsPreset}
                     wysiwygToolbarConfig={wysiwygToolbarConfig}