+++ demo/defaults/args.ts
@@ -17,4 +17,6 @@ export const args: Meta<PlaygroundMiniProps>['args'] = {
     renderPreviewDefined: true,
     height: 'initial',
     directiveSyntax: 'disabled',
+    disabledHTMLBlockModes: [],
+    disableMarkdownItAttrs: true,
 };