+++ demo/stories/presets/Presets.stories.tsx
@@ -1,24 +1,9 @@
 import type {StoryObj} from '@storybook/react';
 
-import {ActionName as Action} from 'src/bundle/config/action-names';
-import {ToolbarName as Toolbar} from 'src/modules/toolbars/constants';
-import {
-    boldItemView,
-    boldItemWysiwyg,
-    colorifyItemMarkup,
-    colorifyItemView,
-    colorifyItemWysiwyg,
-    italicItemMarkup,
-    italicItemView,
-    redoItemMarkup,
-    redoItemView,
-    redoItemWysiwyg,
-    undoItemMarkup,
-    undoItemView,
-    undoItemWysiwyg,
-} from 'src/modules/toolbars/items';
-
 import {Preset as component} from './Preset';
+import {toolbarPresets} from './presets';
+
+const {custom} = toolbarPresets;
 
 export const Zero: StoryObj<typeof component> = {
     args: {preset: 'zero'},
@@ -42,45 +27,7 @@ export const Full: StoryObj<typeof component> = {
 
 export const Custom: StoryObj<typeof component> = {
     args: {
-        toolbarsPreset: {
-            items: {
-                [Action.undo]: {
-                    view: undoItemView,
-                    wysiwyg: undoItemWysiwyg,
-                    markup: undoItemMarkup,
-                },
-                [Action.redo]: {
-                    view: redoItemView,
-                    wysiwyg: redoItemWysiwyg,
-                    markup: redoItemMarkup,
-                },
-                [Action.bold]: {
-                    view: boldItemView,
-                    wysiwyg: boldItemWysiwyg,
-                },
-                [Action.italic]: {
-                    view: italicItemView,
-                    markup: italicItemMarkup,
-                },
-                [Action.colorify]: {
-                    view: colorifyItemView,
-                    wysiwyg: colorifyItemWysiwyg,
-                    markup: colorifyItemMarkup,
-                },
-            },
-            orders: {
-                [Toolbar.wysiwygMain]: [
-                    [Action.colorify],
-                    [Action.bold],
-                    [Action.undo, Action.redo],
-                ],
-                [Toolbar.markupMain]: [
-                    [Action.colorify],
-                    [Action.italic],
-                    [Action.undo, Action.redo],
-                ],
-            },
-        },
+        toolbarsPreset: custom,
     },
 };