+++ src/bundle/settings/index.tsx
@@ -12,6 +12,7 @@ import {
     Menu,
     Popup,
     type PopupPlacement,
+    type QAProps,
 } from '@gravity-ui/uikit';
 
 import {type ClassNameProps, cn} from '../../classname';
@@ -54,6 +55,7 @@ export const EditorSettings = memo<EditorSettingsProps>(function EditorSettings(
                         hotkey="mod+shift+p"
                     >
                         <Button
+                            qa="g-md-markup-preview-button"
                             size="m"
                             view="flat"
                             pin="round-round"
@@ -75,6 +77,7 @@ export const EditorSettings = memo<EditorSettingsProps>(function EditorSettings(
                         pin="round-round"
                         onClick={togglePopup}
                         ref={setChevronElement}
+                        qa="g-md-settings-button"
                         className={bSettings('dropdown-button')}
                     >
                         <Icon data={Gear} />
@@ -87,6 +90,7 @@ export const EditorSettings = memo<EditorSettingsProps>(function EditorSettings(
                     >
                         <SettingsContent
                             {...props}
+                            qa="g-md-settings-content"
                             onClose={hidePopup}
                             className={bSettings('content')}
                         />
@@ -97,18 +101,19 @@ export const EditorSettings = memo<EditorSettingsProps>(function EditorSettings(
     );
 });
 
-type SettingsContentProps = ClassNameProps & {
-    mode: MarkdownEditorMode;
-    onClose: () => void;
-    onModeChange: (mode: MarkdownEditorMode) => void;
-    onShowPreviewChange: (showPreview: boolean) => void;
-    showPreview: boolean;
-    toolbarVisibility: boolean;
-    onToolbarVisibilityChange: (val: boolean) => void;
-    splitMode?: MarkdownEditorSplitMode;
-    splitModeEnabled?: boolean;
-    onSplitModeChange?: (splitModeEnabled: boolean) => void;
-};
+type SettingsContentProps = ClassNameProps &
+    QAProps & {
+        mode: MarkdownEditorMode;
+        onClose: () => void;
+        onModeChange: (mode: MarkdownEditorMode) => void;
+        onShowPreviewChange: (showPreview: boolean) => void;
+        showPreview: boolean;
+        toolbarVisibility: boolean;
+        onToolbarVisibilityChange: (val: boolean) => void;
+        splitMode?: MarkdownEditorSplitMode;
+        splitModeEnabled?: boolean;
+        onSplitModeChange?: (splitModeEnabled: boolean) => void;
+    };
 
 const mdHelpPlacement: PopupPlacement = ['bottom', 'bottom-end', 'right-start', 'right', 'left'];
 
@@ -123,11 +128,13 @@ const SettingsContent: React.FC<SettingsContentProps> = function SettingsContent
     splitModeEnabled,
     className,
     showPreview,
+    qa,
 }) {
     return (
-        <div className={bContent(null, [className])}>
+        <div className={bContent(null, [className])} data-qa={qa}>
             <Menu size="l" className={bContent('mode')}>
                 <Menu.Item
+                    qa="g-md-settings-mode-wysiwyg"
                     active={mode === 'wysiwyg'}
                     onClick={() => {
                         onModeChange('wysiwyg');
@@ -138,6 +145,7 @@ const SettingsContent: React.FC<SettingsContentProps> = function SettingsContent
                     {i18n('settings_wysiwyg')}
                 </Menu.Item>
                 <Menu.Item
+                    qa="g-md-settings-mode-markup"
                     active={mode === 'markup'}
                     onClick={() => {
                         onModeChange('markup');